Fixes for GCC warnings and a work around for what looks like a GCC bug.

+++ cpplox/Scanner/Scanner.h
@@ -39,8 +39,8 @@ class Scanner {
   ErrorReporter &eReporter;
 
   std::list<Token> tokens;
-  int start = 0;
-  int current = 0;
+  size_t start = 0;
+  size_t current = 0;
   int line = 1;
 };
